nfv-auto/requirements.txt
openstacksdk==0.22.0
paramiko==2.4.2
